DOC: improve rst doc display
