+++ altair/schema/_interface/aggregateop.py
@@ -5,7 +5,9 @@ import traitlets as T
 
 
 class AggregateOp(T.Enum):
-    """One of ['values', 'count', 'valid', 'missing', 'distinct', 'sum', 'mean', 'average', 'variance', 'variancep', 'stdev', 'stdevp', 'median', 'q1', 'q3', 'modeskew', 'min', 'max', 'argmin', 'argmax']"""
+    """
+    One of ['values', 'count', 'valid', 'missing', 'distinct', 'sum', 'mean', 'average', 'variance', 'variancep', 'stdev', 'stdevp', 'median', 'q1', 'q3', 'modeskew', 'min', 'max', 'argmin', 'argmax']
+    """
     def __init__(self, default_value=T.Undefined, **metadata):
         super(AggregateOp, self).__init__(['values', 'count', 'valid', 'missing', 'distinct', 'sum', 'mean', 'average', 'variance', 'variancep', 'stdev', 'stdevp', 'median', 'q1', 'q3', 'modeskew', 'min', 'max', 'argmin', 'argmax'],
                                     default_value=default_value,
