+++ altair/schema/_interface/axisorient.py
@@ -5,7 +5,9 @@ import traitlets as T
 
 
 class AxisOrient(T.Enum):
-    """One of ['top', 'right', 'left', 'bottom']"""
+    """
+    One of ['top', 'right', 'left', 'bottom']
+    """
     def __init__(self, default_value=T.Undefined, **metadata):
         super(AxisOrient, self).__init__(['top', 'right', 'left', 'bottom'],
                                     default_value=default_value,
