+++ altair/schema/_interface/dataformat.py
@@ -5,7 +5,9 @@ import traitlets as T
 
 
 class DataFormat(T.Enum):
-    """One of ['json', 'csv', 'tsv']"""
+    """
+    One of ['json', 'csv', 'tsv']
+    """
     def __init__(self, default_value=T.Undefined, **metadata):
         super(DataFormat, self).__init__(['json', 'csv', 'tsv'],
                                     default_value=default_value,
