+++ altair/schema/_interface/fontweight.py
@@ -5,7 +5,9 @@ import traitlets as T
 
 
 class FontWeight(T.Enum):
-    """One of ['normal', 'bold']"""
+    """
+    One of ['normal', 'bold']
+    """
     def __init__(self, default_value=T.Undefined, **metadata):
         super(FontWeight, self).__init__(['normal', 'bold'],
                                     default_value=default_value,
