+++ altair/schema/_interface/mark.py
@@ -5,7 +5,9 @@ import traitlets as T
 
 
 class Mark(T.Enum):
-    """One of ['area', 'bar', 'line', 'point', 'text', 'tick', 'rule', 'circle', 'square']"""
+    """
+    One of ['area', 'bar', 'line', 'point', 'text', 'tick', 'rule', 'circle', 'square']
+    """
     def __init__(self, default_value=T.Undefined, **metadata):
         super(Mark, self).__init__(['area', 'bar', 'line', 'point', 'text', 'tick', 'rule', 'circle', 'square'],
                                     default_value=default_value,
