+++ altair/schema/_interface/verticalalign.py
@@ -5,7 +5,9 @@ import traitlets as T
 
 
 class VerticalAlign(T.Enum):
-    """One of ['top', 'middle', 'bottom']"""
+    """
+    One of ['top', 'middle', 'bottom']
+    """
     def __init__(self, default_value=T.Undefined, **metadata):
         super(VerticalAlign, self).__init__(['top', 'middle', 'bottom'],
                                     default_value=default_value,
