+++ doc/API.rst
@@ -10,6 +10,43 @@ Top-Level Objects
 
 .. altair-class:: altair.FacetedChart
 
+Encoding Collections
+--------------------
+
+.. altair-class:: altair.Encoding
+
+.. altair-class:: altair.Facet
+
+Encoding Channels
+-----------------
+
+.. altair-class:: altair.Color
+
+.. altair-class:: altair.Column
+
+.. altair-class:: altair.Detail
+
+.. altair-class:: altair.Label
+
+.. altair-class:: altair.Opacity
+
+.. altair-class:: altair.Order
+
+.. altair-class:: altair.Path
+
+.. altair-class:: altair.Row
+
+.. altair-class:: altair.Shape
+
+.. altair-class:: altair.Size
+
+.. altair-class:: altair.Text
+
+.. altair-class:: altair.X
+
+.. altair-class:: altair.Y
+
+
 Other Objects
 -------------
 
@@ -17,4 +54,4 @@ Other Objects
    :members:
    :imported-members:
    :undoc-members:
-   :exclude-members: Chart,LayeredChart,FacetedChart
+   :exclude-members: Chart,LayeredChart,FacetedChart,Encoding,Facet,Color,Column,Detail,Label,Opacity,Order,Path,Row,Shape,Size,Text,X,Y
