+++ tools/templates/interface_enum.tpl
@@ -5,7 +5,9 @@ import traitlets as T
 
 
 class {{ cls.name }}(T.Enum):
-    """One of [{{ cls.enum|map('repr')|join(', ') }}]"""
+    """
+    One of [{{ cls.enum|map('repr')|join(', ') }}]
+    """
     def __init__(self, default_value=T.Undefined, **metadata):
         super({{ cls.name }}, self).__init__([{{ cls.enum|map('repr')|join(', ') }}],
                                     default_value=default_value,
